Relax validation on dependencies in build statements
In the validation code for the build statement class, I used to throw
errors if the list of order-only or implicit dependencies was
empty. While this is conceptually correct, the dependency information
will only get appended to the build rule if the list is not empty, so
we actually don't need to have such aggressive validation. This saves
the generator code from having to check the size of the dependency
lists when adding a build statement.